Disconnect Nodes with old 10070001 Protocol Version.
That number was set by the previous Riecoin Developer for some reason, and then we put back Bitcoin's 70016 since 22.03.
Gets rid of connections from nodes that are somehow still Pre Fork 2...
Affects 0.21 as well, but they are now a small minority and will still have some 23.04- Nodes to fetch Blocks from for a while...

Also simplify a bit the logic (there are no Riecoin Nodes with Protocol Version < 70016).

Signed-off-by: Pttn <28868425+Pttn@users.noreply.github.com>